Move initializing ClusterManager into a runable
