+++ src/main/catalog-sources/kubeconfig-sync/compute-diff.injectable.ts
@@ -12,8 +12,8 @@ import type { CatalogEntity } from "../../../common/catalog";
 import getClusterByIdInjectable from "../../../common/cluster-store/get-by-id.injectable";
 import type { Cluster } from "../../../common/cluster/cluster";
 import { loadConfigFromString } from "../../../common/kube-helpers";
-import { catalogEntityFromCluster } from "../../cluster-manager";
-import clusterManagerInjectable from "../../cluster-manager.injectable";
+import { catalogEntityFromCluster } from "../../cluster/manager";
+import clusterManagerInjectable from "../../cluster/manager.injectable";
 import createClusterInjectable from "../../create-cluster/create-cluster.injectable";
 import configToModelsInjectable from "./config-to-models.injectable";
 import kubeconfigSyncLoggerInjectable from "./logger.injectable";
