+++ src/main/cluster/manager.ts
@@ -3,20 +3,20 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 
-import "../common/ipc/cluster";
+import "../../common/ipc/cluster";
 import type http from "http";
 import { action, makeObservable, observable, observe, reaction, toJS } from "mobx";
-import type { Cluster } from "../common/cluster/cluster";
-import logger from "./logger";
-import { apiKubePrefix } from "../common/vars";
-import { getClusterIdFromHost, isErrnoException } from "../common/utils";
-import type { KubernetesClusterPrometheusMetrics } from "../common/catalog-entities/kubernetes-cluster";
-import { isKubernetesCluster, KubernetesCluster, LensKubernetesClusterStatus } from "../common/catalog-entities/kubernetes-cluster";
-import { ipcMainOn } from "../common/ipc";
+import type { Cluster } from "../../common/cluster/cluster";
+import logger from "../logger";
+import { apiKubePrefix } from "../../common/vars";
+import { getClusterIdFromHost, isErrnoException } from "../../common/utils";
+import type { KubernetesClusterPrometheusMetrics } from "../../common/catalog-entities/kubernetes-cluster";
+import { isKubernetesCluster, KubernetesCluster, LensKubernetesClusterStatus } from "../../common/catalog-entities/kubernetes-cluster";
+import { ipcMainOn } from "../../common/ipc";
 import { once } from "lodash";
-import type { ClusterStore } from "../common/cluster-store/cluster-store";
-import type { ClusterId } from "../common/cluster-types";
-import type { CatalogEntityRegistry } from "./catalog";
+import type { ClusterStore } from "../../common/cluster-store/cluster-store";
+import type { ClusterId } from "../../common/cluster-types";
+import type { CatalogEntityRegistry } from "../catalog";
 
 const logPrefix = "[CLUSTER-MANAGER]:";
