+++ src/main/lens-proxy/proxy-functions/shell-api-request/shell-api-request.injectable.ts
@@ -6,7 +6,7 @@ import { getInjectable } from "@ogre-tools/injectable";
 import { shellApiRequest } from "./shell-api-request";
 import createShellSessionInjectable from "../../../shell-session/create-shell-session.injectable";
 import shellRequestAuthenticatorInjectable from "./shell-request-authenticator/shell-request-authenticator.injectable";
-import clusterManagerInjectable from "../../../cluster-manager.injectable";
+import clusterManagerInjectable from "../../../cluster/manager.injectable";
 
 const shellApiRequestInjectable = getInjectable({
   id: "shell-api-request",
